Feature/successful modal (#1680)
* Add modal component and enhance checkout UI

Introduced a new `Modal` component to be used in the checkout process for displaying detailed views. Enhanced `Checkout.vue` and `Result.vue` to integrate the modal functionality, allowing better visualization of certain elements. Added Japanese translations for "Passport" and "Home".

* Refactor modal logic and remove unused image

Renamed the function responsible for opening the modal to 'modalOpen' for clarity. Removed the initialization and usage of the unused 'img2' image. Ensured the modal opens immediately after image loading, and updated the content binding of the modal to use dynamic attributes.